# apps/api/openapi/admin.json
{"openapi":"3.1.0","info":{"title":"DrunkGraph Admin API","description":"Backoffice API for managing places, drinks, and flavors.\n\n## How to get a token\n\nAdmin tokens are issued by the **admin tenant** in FusionAuth — a separate tenant from regular users. JWTs from the user tenant will be rejected here.\n\n1. Log in to the admin backoffice at [http://localhost:3000/admin](http://localhost:3000/admin)\n2. Retrieve your token from the session (via `/api/proxy/token/debug` on the admin app)\n3. Click the **Authorize** lock icon above and paste it\n\n## Batch drink import\n\nThe typical workflow for adding drinks to a place:\n\n1. `POST /api/admin/uploads/sign` — get a Cloudinary signature\n2. Upload images directly to Cloudinary from the browser using the signed params\n3. Collect the returned `public_id` values\n4. `POST /api/admin/places/{placeId}/drinks/batch` — submit the batch with image public IDs\n\nThe server never handles image binary data.\n","version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/admin/flavors/{name}":{"put":{"tags":["admin-flavor-handler"],"operationId":"updateFlavor","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlavorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flavor"}}}}}},"delete":{"tags":["admin-flavor-handler"],"operationId":"deleteFlavor","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/admin/drinks/{id}":{"get":{"tags":["admin-drink-handler"],"operationId":"adminGetDrink","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Drink"}}}}}},"put":{"tags":["admin-drink-handler"],"operationId":"updateDrink","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrinkEditRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Drink"}}}}}},"delete":{"tags":["admin-drink-handler"],"operationId":"deleteDrink","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/admin/uploads/sign":{"post":{"tags":["admin-upload-handler"],"operationId":"signUpload","parameters":[{"name":"folder","in":"query","required":false,"schema":{"type":"string","default":"drinks"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/admin/places":{"get":{"tags":["admin-place-handler"],"operationId":"listPlaces","parameters":[{"name":"search","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Place"}}}}}}},"post":{"tags":["admin-place-handler"],"operationId":"createPlace","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Place"}}}}}}},"/api/admin/places/{placeId}/drinks/batch":{"post":{"tags":["admin-drink-handler"],"operationId":"importDrinks","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DrinkBatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Drink"}}}}}}}},"/api/admin/flavors":{"get":{"tags":["admin-flavor-handler"],"operationId":"adminListFlavors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Flavor"}}}}}}},"post":{"tags":["admin-flavor-handler"],"operationId":"createFlavor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlavorRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Flavor"}}}}}}},"/api/admin/drinks":{"get":{"tags":["admin-drink-handler"],"operationId":"adminListDrinks","parameters":[{"name":"search","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Drink"}}}}}}}}},"components":{"schemas":{"FlavorRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["name"]},"Flavor":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"DrinkEditRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"category":{"type":"string","minLength":1},"placeId":{"type":"string","minLength":1},"alcoholPct":{"type":"number","format":"double"},"price":{"type":"number","format":"double"},"flavors":{"type":"object","additionalProperties":{"type":"number","format":"double"}},"imagePublicIds":{"type":"array","items":{"type":"string"}}},"required":["category","name","placeId"]},"Drink":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"placeId":{"type":"string"},"alcoholPct":{"type":"number","format":"double"},"price":{"type":"number","format":"double"},"flavors":{"type":"object","additionalProperties":{"type":"number","format":"double"}},"imageUrls":{"type":"array","items":{"type":"string"}}}},"PlaceRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"location":{"type":"string","minLength":1}},"required":["location","name"]},"Place":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"location":{"type":"string"}}},"DrinkBatchRequest":{"type":"object","properties":{"drinks":{"type":"array","items":{"$ref":"#/components/schemas/DrinkItemRequest"},"minItems":1}},"required":["drinks"]},"DrinkItemRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"category":{"type":"string","minLength":1},"alcoholPct":{"type":"number","format":"double"},"price":{"type":"number","format":"double"},"flavors":{"type":"object","additionalProperties":{"type":"number","format":"double"}},"imagePublicIds":{"type":"array","items":{"type":"string"}}},"required":["category","name"]}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}